[FEATURE] New goal for moths that will draw them to flames.
Made butterfly construction smoother using on-access creation of the butterfly data reference. Modified goals so butterflies exit the landed state when a goal ends instead of relying on other goals to do this for them. Modified the wander goal so it now uses Minecraft's superior pathfinding code with an interval of `1` so they don't hover in place.

The new `MothWanderGoal` will cause moths to "reroll" their target position if the first one is darker. This will cause them to tend toward brighter positions, but won't prevent them from moving to darker block positions.