[FEATURE] Added the Domestic Silk Moth.
Also includes some minor updates to localisation string generation, and fixes butterfly eggs getting data at the wrong times. It turns out that using underscores in IDs confuses the butterfly data code, so there is a hack to fix it for domestic silk moths, but in the future underscores shouldn't be used in butterfly IDs.